Web parsing code; clean up which primitives we import. Add scripting for merge/commit into db

+++ generate_database.sh
@@ -1,28 +1,62 @@
 #!/bin/bash -i
 
+set -o allexport
+
 . user_environment.sh
+
+fuzz=false
+merge=false
+git_commit=false
+
+while getopts 'fgm' flag; do
+  case "${flag}" in
+    f) fuzz=true ;;
+    m) merge=true ;;
+    g) git_commit=true ;;
+    *) print_usage
+       exit 1 ;;
+  esac
+done
 
 pushd tools
 #python3 tilegrid_all.py
 popd
 
 PRJOXIDE_ROOT=`pwd`
-pushd fuzzers
 
 run_fuzzer() {
     dir="$1"
     if [ -f "$dir/fuzzer.py" ]; then
+        set -ex
+
         echo "=================== Entering $dir ==================="
         pushd "$dir" > /dev/null || return
-        rm -rf db .deltas
-        $PRJOXIDE_ROOT/link-db-root.sh
-        FUZZER_TITLE=$dir PRJOXIDE_DB="$(pwd)/db" python3 fuzzer.py 2>&1 | tee >(gzip --stdout > fuzzer.log.gz)
+
+        if [ "$fuzz" = true ] ; then
+          rm -rf db .deltas
+          $PRJOXIDE_ROOT/link-db-root.sh
+          FUZZER_TITLE=$dir PRJOXIDE_DB="$(pwd)/db" python3 fuzzer.py 2>&1 | tee >(gzip --stdout > fuzzer.log.gz)
+        fi
         popd > /dev/null || true
+
+        if [ -d "$dir/db" ]; then
+          if [ "$merge" = true ] ; then
+            pushd ..
+            python3 ./tools/merge-databases.py fuzzers/$dir/db database/
+            popd
+          fi
+
+          if [ "$git_commit" = true ] ; then
+            pushd ../database
+            git add **.ron
+            git commit -m "Incorporating database changes from $dir"
+            popd
+          fi
+        fi
+
     fi
 }
-
 export -f run_fuzzer
-export PRJOXIDE_ROOT
 
-#find . -mindepth 1 -maxdepth 1 -type d ! -name LIFCL -exec bash -c 'run_fuzzer "$0"' {} \;
-find LIFCL -mindepth 1 -maxdepth 1 -type d | sort | xargs -I {} bash -c 'run_fuzzer "{}"'
+pushd fuzzers
+find . -mindepth 1 -maxdepth 2 -type d | sort | xargs -I {} bash -c 'run_fuzzer "{}"'
